refactor(asset): rename controllers and routes to asset
