+++ src/controllers/index.ts
@@ -1,3 +1,3 @@
+export * from './asset.ts';
 export * from './record.ts';
-export * from './systemRecord.ts';
 export * from './validate.ts';
