+++ tests/unit/services/asset.test.ts
